msft warnings

+++ include/phase4/engine/common/bitset.h
@@ -161,7 +161,7 @@ inline constexpr Bitset Bitset::MAX(0b11111111'11111111'11111111'11111111'111111
 	// GCC or Clang
 	return __builtin_popcountll(m_bits);
 #elif defined(_MSC_VER)
-	return __popcnt64(m_bits);
+	return static_cast<uint8_t>(__popcnt64(m_bits));
 #else
 	static_assert(false, "Define USE_SLOW_BITSET_COUNT to use internal implementation.")
 #endif
@@ -170,7 +170,7 @@ inline constexpr Bitset Bitset::MAX(0b11111111'11111111'11111111'11111111'111111
 [[nodiscard]] inline constexpr uint8_t Bitset::count() const noexcept {
 #if defined(__GNUC__) || defined(__clang__)
 	// GCC or Clang
-	return static_cast<std::size_t>(__builtin_popcountll(m_bits));
+	return static_cast<uint8_t>(__builtin_popcountll(m_bits));
 #elif defined(_MSC_VER) && 0 // Skip intrinsic for Microsoft Visual C++ due to not being constexpr
 #else
 	// Fallback implementation for other compilers or platforms
